requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.13
# by the following command:
#
#    pip-compile --cert=None --client-cert=None --index-url=None --pip-args=None
#
aiohappyeyeballs==2.6.1
    # via aiohttp
aiohttp==3.12.15
    # via langchain-community
aiosignal==1.4.0
    # via aiohttp
annotated-types==0.7.0
    # via pydantic
anyio==4.11.0
    # via
    #   httpx
    #   starlette
    #   watchfiles
attrs==25.3.0
    # via aiohttp
brotli==1.1.0
    # via httpx
certifi==2025.8.3
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.4.3
    # via requests
click==8.3.0
    # via
    #   ddgs
    #   duckduckgo-search
    #   uvicorn
colorama==0.4.6
    # via
    #   click
    #   uvicorn
dataclasses-json==0.6.7
    # via langchain-community
ddgs==9.6.0
    # via -r requirements.in
duckduckgo-search==8.1.1
    # via -r requirements.in
fastapi==0.118.0
    # via -r requirements.in
frozenlist==1.7.0
    # via
    #   aiohttp
    #   aiosignal
greenlet==3.2.4
    # via sqlalchemy
h11==0.16.0
    # via
    #   httpcore
    #   uvicorn
h2==4.3.0
    # via httpx
hpack==4.1.0
    # via h2
httpcore==1.0.9
    # via httpx
httptools==0.6.4
    # via uvicorn
httpx[brotli,http2,socks]==0.28.1
    # via
    #   ddgs
    #   langsmith
    #   ollama
httpx-sse==0.4.1
    # via langchain-community
hyperframe==6.1.0
    # via h2
idna==3.10
    # via
    #   anyio
    #   httpx
    #   requests
    #   yarl
jsonpatch==1.33
    # via langchain-core
jsonpointer==3.0.0
    # via jsonpatch
langchain==0.3.27
    # via
    #   -r requirements.in
    #   langchain-community
langchain-community==0.3.30
    # via -r requirements.in
langchain-core==0.3.77
    # via
    #   langchain
    #   langchain-community
    #   langchain-ollama
    #   langchain-text-splitters
langchain-ollama==0.3.9
    # via -r requirements.in
langchain-text-splitters==0.3.11
    # via langchain
langsmith==0.4.31
    # via
    #   langchain
    #   langchain-community
    #   langchain-core
lxml==6.0.2
    # via
    #   ddgs
    #   duckduckgo-search
marshmallow==3.26.1
    # via dataclasses-json
multidict==6.6.4
    # via
    #   aiohttp
    #   yarl
mypy-extensions==1.1.0
    # via typing-inspect
numpy==2.3.3
    # via langchain-community
ollama==0.6.0
    # via langchain-ollama
orjson==3.11.3
    # via langsmith
packaging==25.0
    # via
    #   langchain-core
    #   langsmith
    #   marshmallow
primp==0.15.0
    # via
    #   ddgs
    #   duckduckgo-search
propcache==0.3.2
    # via
    #   aiohttp
    #   yarl
pydantic==2.11.9
    # via
    #   fastapi
    #   langchain
    #   langchain-core
    #   langsmith
    #   ollama
    #   pydantic-settings
pydantic-core==2.33.2
    # via pydantic
pydantic-settings==2.11.0
    # via langchain-community
python-dotenv==1.1.1
    # via
    #   -r requirements.in
    #   pydantic-settings
    #   uvicorn
pyyaml==6.0.3
    # via
    #   langchain
    #   langchain-community
    #   langchain-core
    #   uvicorn
requests==2.32.5
    # via
    #   langchain
    #   langchain-community
    #   langsmith
    #   requests-toolbelt
requests-toolbelt==1.0.0
    # via langsmith
sniffio==1.3.1
    # via anyio
socksio==1.0.0
    # via httpx
sqlalchemy==2.0.43
    # via
    #   langchain
    #   langchain-community
starlette==0.48.0
    # via fastapi
tenacity==9.1.2
    # via
    #   langchain-community
    #   langchain-core
typing-extensions==4.15.0
    # via
    #   fastapi
    #   langchain-core
    #   pydantic
    #   pydantic-core
    #   sqlalchemy
    #   typing-inspect
    #   typing-inspection
typing-inspect==0.9.0
    # via dataclasses-json
typing-inspection==0.4.2
    # via
    #   pydantic
    #   pydantic-settings
urllib3==2.5.0
    # via requests
uvicorn[standard]==0.37.0
    # via -r requirements.in
watchfiles==1.1.0
    # via uvicorn
websockets==15.0.1
    # via uvicorn
yarl==1.20.1
    # via aiohttp
zstandard==0.25.0
    # via langsmith
